Added alternative framework via xwdMaster - test with puzzle21.xwd.html
xwdMaster is a single-include script which then includes other needed
js and css files. It allows the puzzle to be presented in a .xwd.html
file which is an only slightly modified version of the plain text
crossword file. (But would still be auto-generated, like .xwd.js files.)
Other than a small change to application.js (to call the crossword reader
at "runtime") all the same .js files are used, omitting xwdHtml.
Probably a dead-end experiment - I think there's a trick with <iframe>
that will let me include and then read a plain text file.