Made local storage specific to puzzle and refined cursor movement
Puzzle name is now added to "Xwd-" prefix for label for local storage.
This means grid is correctly redrawn when a different puzzle is
loaded. Cursor move now decides on which spot to be in (in new cell)
even if the cell is in the same spot as before. (Before it wasn't
letting you use the mouse function to change spots in same cell.)