Switched to event manager which doesn't rely on function.bind()
The old code was needlessly confusing - functions which were methods
of the (unique) game manager were being bound to that game manager
in order to be registered with the event listener, which in any case
belonged to the game manager. Now the event listenere registers
[ target, function ] pairs, so binding is unnecessary.
Also, the event manager is now in xwdEventManager.js and the main
game manager is in xwdInterface2.js.